Update dependencies, downgrade detekt dependency and fix breaking changes

+++ src/main/kotlin/de/manuzid/staticcodereviewplugin/model/GitLabConfiguration.kt
@@ -17,7 +17,7 @@
 package de.manuzid.staticcodereviewplugin.model
 
 data class GitLabConfiguration(val gitLabUrl: String, val authentication: GitLabAuthenticationConfiguration,
-                               val projectId: String, val mergeRequestIid: Int, val proxyConfiguration: ProxyConfiguration? = null)
+                               val projectId: String, val mergeRequestIid: Long, val proxyConfiguration: ProxyConfiguration? = null)
 
 data class GitLabAuthenticationConfiguration(val token: String?, val username: String?, val password: String?)
